feat: split python app and startup logic
We install the startup scripts in /startup instead of in the /app
directory - this way we can mount over /app and still have all the
startup logic accessible.

The `wsgi.py` now also resides in the /startup directory and we pass the
--pythonpath argument to gunicorn to find it.

Additionaly, we install a /usr/local/bin/apis script, that runs the
scripts in the /startup folder.